combine the activator into the existing service impl.
